+++ frontend/src/lib/components/assets/AssetGraph/PipelineGraphEditor.svelte
@@ -1,7 +1,12 @@
 <script lang="ts">
-	import type { Snippet } from 'svelte'
+	import { onMount, untrack, type Snippet } from 'svelte'
 	import { Loader2 } from 'lucide-svelte'
 	import { Pane, Splitpanes } from 'svelte-splitpanes'
+	import { DraftService } from '$lib/gen'
+	import { decodeState, encodeState } from '$lib/utils'
+	import { UserDraftDbSyncer } from '$lib/userDraftDbSyncer.svelte'
+	import { extractWrites } from '$lib/components/assets/lib'
+	import type { PipelineDraft } from './pipelineAiHelpers'
 	import HideButton from '$lib/components/apps/editor/settingsPanel/HideButton.svelte'
 	import AssetGraphCanvas from './AssetGraphCanvas.svelte'
 	import AssetGraphDetailsPane from './AssetGraphDetailsPane.svelte'
@@ -25,6 +30,8 @@
 		displayGraph,
 		mode,
 		workspace,
+		folder,
+		persistDrafts = false,
 		pathPrefix,
 		defaultPathSuffix = 'new_pipeline_script',
 		panelHidden = false,
@@ -83,6 +90,11 @@
 		displayGraph: AssetGraphResponse
 		mode: PipelineMode
 		workspace: string | undefined
+		/** Folder the pipeline is scoped to — drives the autosave draft path. */
+		folder: string
+		/** Persist drafts as a `data_pipeline` DraftService bundle (route page only;
+		 * the session preview leaves this false). FlowBuilder's autosave analogue. */
+		persistDrafts?: boolean
 		pathPrefix: string
 		defaultPathSuffix?: string
 		panelHidden?: boolean
@@ -168,15 +180,180 @@
 		mode !== 'edit' && editor.selection == undefined && editor.activeDraftPath == undefined
 	)
 
+	// Wrap the writes (and the rightPaneSize read in the else branch) in untrack so
+	// the effect tracks only `detailsPaneOpen` / `storedRightPaneSize` — without it,
+	// the Pane `bind:size` feedback loops the effect and pegs the main thread.
 	$effect(() => {
 		if (detailsPaneOpen) {
-			const restore = storedRightPaneSize > 0 ? storedRightPaneSize : 40
-			rightPaneSize = restore
-			leftPaneSize = 100 - restore
+			const restore = storedRightPaneSize
+			untrack(() => {
+				rightPaneSize = restore > 0 ? restore : 40
+				leftPaneSize = 100 - (restore > 0 ? restore : 40)
+			})
 		} else {
-			if (rightPaneSize > 0) storedRightPaneSize = rightPaneSize
-			leftPaneSize = 100
+			untrack(() => {
+				if (rightPaneSize > 0) storedRightPaneSize = rightPaneSize
+				leftPaneSize = 100
+			})
 		}
+	})
+
+	// ===================== Draft autosave (persistDrafts only) =====================
+	// All of this folder's in-flight drafts live in ONE per-user DB draft (typ
+	// `data_pipeline`) keyed at the folder, syncing across devices + surfacing in
+	// the global drafts list. localStorage is a synchronous crash mirror, READ only
+	// for the one-time migration below; the DB is the source of truth on load.
+	// FlowBuilder's autosave analogue — gated by `persistDrafts` so the in-session
+	// preview (which doesn't persist) opts out.
+	const PIPELINE_DRAFT_KIND = 'data_pipeline' as const
+	let pipelineDraftPath = $derived(`f/${folder}/data_pipeline`)
+	let storageKey = $derived(`pipeline-${folder}`)
+	type PipelineDraftBundle = { drafts: Array<[string, PipelineDraft]>; activeDraftPath?: string }
+	let draftsHydrated = $state(false)
+	let lastPersistedBundle: string | undefined = undefined
+
+	function restoreBundle(bundle: PipelineDraftBundle) {
+		if (Array.isArray(bundle.drafts)) {
+			const loaded = new Map<string, PipelineDraft>()
+			for (const entry of bundle.drafts) {
+				if (entry && typeof entry[0] === 'string' && entry[1]?.script) {
+					const d = entry[1] as PipelineDraft
+					if (typeof d.localId !== 'string' || d.localId === '')
+						d.localId = editor.newDraftLocalId()
+					loaded.set(entry[0], d)
+				}
+			}
+			if (loaded.size > 0) editor.drafts = loaded
+		}
+		if (typeof bundle.activeDraftPath === 'string') editor.activeDraftPath = bundle.activeDraftPath
+	}
+
+	function readLocalBundle(): PipelineDraftBundle | undefined {
+		if (typeof localStorage === 'undefined') return undefined
+		const raw = localStorage.getItem(`pipeline-${folder}`)
+		if (!raw) return undefined
+		try {
+			const s = decodeState(raw)
+			if (s && (Array.isArray(s.drafts) || typeof s.activeDraftPath === 'string')) {
+				return {
+					drafts: Array.isArray(s.drafts) ? s.drafts : [],
+					activeDraftPath: typeof s.activeDraftPath === 'string' ? s.activeDraftPath : undefined
+				}
+			}
+		} catch (e) {
+			console.warn('failed to read local pipeline state', e)
+		}
+		return undefined
+	}
+
+	async function hydrateDrafts() {
+		const ws = workspace
+		const path = pipelineDraftPath
+		try {
+			let bundle: PipelineDraftBundle | undefined
+			let serverSavedAt: string | undefined
+			if (ws) {
+				const row = await DraftService.getOwnDraft({
+					workspace: ws,
+					kind: PIPELINE_DRAFT_KIND,
+					path
+				})
+				if (row?.value) {
+					bundle = row.value as PipelineDraftBundle
+					serverSavedAt = row.created_at
+					editor.loadedFromDbDraft = true
+				}
+			}
+			let migratedFromLocal = false
+			if (!bundle) {
+				const local = readLocalBundle()
+				if (local) {
+					bundle = local
+					migratedFromLocal = true
+				}
+			}
+			if (bundle) restoreBundle(bundle)
+			UserDraftDbSyncer.recordRemoteSync(
+				{ workspace: ws ?? '', itemKind: PIPELINE_DRAFT_KIND, path },
+				migratedFromLocal ? undefined : serverSavedAt
+			)
+			if (!migratedFromLocal) lastPersistedBundle = bundle ? JSON.stringify(bundle) : undefined
+		} catch (e) {
+			console.warn('failed to load pipeline drafts', e)
+		} finally {
+			draftsHydrated = true
+		}
+	}
+
+	onMount(() => {
+		if (persistDrafts) void hydrateDrafts()
+	})
+
+	$effect(() => {
+		if (!persistDrafts) return
+		// For the active draft, snapshot the latest live body writes + editor buffer
+		// at serialize time so edits since the last pane-transition aren't lost on reload.
+		const liveWritesSnapshot =
+			editor.liveBodyAssets.scriptPath != undefined &&
+			editor.drafts.has(editor.liveBodyAssets.scriptPath)
+				? extractWrites(editor.liveBodyAssets.assets)
+				: undefined
+		const liveWritesPath = editor.liveBodyAssets.scriptPath
+		const liveContentPath =
+			editor.liveContent.scriptPath != undefined && editor.drafts.has(editor.liveContent.scriptPath)
+				? editor.liveContent.scriptPath
+				: undefined
+		const liveContentValue = editor.liveContent.content
+		const serialized = Array.from(editor.drafts.entries()).map(([p, d]) => {
+			const outputAssets =
+				liveWritesSnapshot != undefined && liveWritesPath === p
+					? liveWritesSnapshot.length > 0
+						? liveWritesSnapshot
+						: undefined
+					: d.outputAssets
+			const script =
+				liveContentPath === p && d.script.content !== liveContentValue
+					? { ...d.script, content: liveContentValue }
+					: d.script
+			if (script === d.script && outputAssets === d.outputAssets)
+				return [p, d] as [string, PipelineDraft]
+			return [p, { ...d, script, outputAssets }] as [string, PipelineDraft]
+		})
+		const activePath = editor.activeDraftPath
+		const key = storageKey
+		const ws = workspace
+		const path = pipelineDraftPath
+		const hydrated = draftsHydrated
+		untrack(() => {
+			if (!hydrated) return
+			const isEmpty = serialized.length === 0 && !activePath
+			const bundle: PipelineDraftBundle | undefined = isEmpty
+				? undefined
+				: { drafts: serialized, activeDraftPath: activePath }
+			try {
+				if (typeof localStorage !== 'undefined') {
+					if (isEmpty) localStorage.removeItem(key)
+					else
+						localStorage.setItem(
+							key,
+							encodeState({ drafts: serialized, activeDraftPath: activePath })
+						)
+				}
+			} catch (e) {
+				console.warn('failed to mirror pipeline state', e)
+			}
+			const serializedBundle = bundle ? JSON.stringify(bundle) : undefined
+			if (serializedBundle === lastPersistedBundle) return
+			lastPersistedBundle = serializedBundle
+			if (!ws) return
+			void UserDraftDbSyncer.save({
+				workspace: ws,
+				itemKind: PIPELINE_DRAFT_KIND,
+				path,
+				value: bundle ?? null,
+				auto: true
+			})
+		})
 	})
 </script>
